requirements.txt
openai
pillow
pyqt5